# api/protocol/api/openapi.yaml
openapi: 3.0.0
info:
  title: Index and Embedding API
  version: 1.0.0
servers:
- url: /
paths:
...
  /embedding:
    get:
      description: Retrieves a list of all Embedding entities.
      operationId: listEmbeddings
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Embedding'
                type: array
          description: A list of Embeddings
        "401":
          description: Unauthorized. Authentication required to access this resource.
      summary: List all Embeddings
      x-eov-operation-handler: controllers/DefaultController
    post:
      description: Endpoint to create a new Embedding entity.
      operationId: createEmbedding
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Embedding'
        required: true
      responses:
        "201":
          description: Embedding created successfully
        "400":
          description: Invalid input
        "401":
          description: Unauthorized. Authentication required to access this resource.
      security:
      - bearerAuth: []
      summary: Create a new Embedding
      x-eov-operation-handler: controllers/DefaultController
  /embedding/{embeddingId}:
    get:
      description: Retrieve an Embedding entity by its unique identifier.
      operationId: getEmbedding
      parameters:
      - explode: false
        in: path
        name: embeddingId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Embedding'
          description: Successful response
        "401":
          description: Unauthorized. Authentication required to access this resource.
        "404":
          description: Embedding not found
      summary: Get an Embedding by ID
      x-eov-operation-handler: controllers/DefaultController
    put:
      description: Update an existing Embedding entity by its ID.
      operationId: updateEmbedding
      parameters:
      - explode: false
        in: path
        name: embeddingId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Embedding'
        required: true
      responses:
        "200":
          description: Embedding updated successfully
        "400":
          description: Invalid input
        "401":
          description: Unauthorized. Authentication required to access this resource.
        "404":
          description: Embedding not found
      security:
      - bearerAuth: []
      summary: Update an Embedding by ID
      x-eov-operation-handler: controllers/DefaultController
...
components:
  schemas:
...
    Embedding:
      example:
        modelName: modelName
        itemId: itemId
        createdAt: 2000-01-23T04:56:07.000+00:00
        deletedAt: 2000-01-23T04:56:07.000+00:00
        context:
          context: context
          contextDescription: contextDescription
          category: category
        indexId: indexId
        controllerDID: controllerDID
        vector:
        - 0.8008281904610115
        - 0.8008281904610115
        version: version
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        modelName:
          type: string
        vector:
          items:
            type: number
          type: array
        context:
          $ref: '#/components/schemas/EmbeddingContext'
        indexId:
          type: string
        itemId:
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        deletedAt:
          format: date-time
          type: string
        controllerDID:
          type: string
        version:
          type: string
      type: object
    EmbeddingContext:
      example:
        context: context
        contextDescription: contextDescription
        category: category
      properties:
        context:
          type: string
        contextDescription:
          type: string
        category:
          type: string
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: EIP-4361 Signature
      scheme: bearer
      type: http